+++ src/Resxible/Localization/ResourceFileHandlerBase.cs
@@ -11,13 +11,11 @@ namespace Com.Apcurium.Resxible.Localization
     {
         private readonly string _filePath;
         private readonly HashSet<string> _duplicateKeys;
-        private readonly IList<HtmlEntity> _htmlEntities = new List<HtmlEntity>();
-
+        
         protected ResourceFileHandlerBase(string filePath) : base(StringComparer.OrdinalIgnoreCase)
         {
             _duplicateKeys = new HashSet<string>();
             _filePath = filePath;
-            LoadHtmlEntities();
             CreateFileIfMissing ();
         }
 
@@ -25,28 +23,7 @@ namespace Com.Apcurium.Resxible.Localization
         {
             get { return Path.GetFileName(_filePath); }
         }
-
-        protected string DecodeXml(string text)
-        {
-            //Others invalid characters does not look to be escaped...
-            foreach (var htmlEntity in _htmlEntities)
-            {
-                text = text.Replace(htmlEntity.Encoded, htmlEntity.Normal);
-            }
-
-            return text;
-        }
-
-        protected string EncodeXml(string text)
-        {
-            foreach (var htmlEntity in _htmlEntities)
-            {
-                text = text.Replace(htmlEntity.Normal, htmlEntity.Encoded);
-            }
-
-            return text;
-        }
-
+        
         protected void TryAdd(string key, string value)
         {
             if (ContainsKey(key))
@@ -75,6 +52,16 @@ namespace Com.Apcurium.Resxible.Localization
             return backupFilePath;
         }
 
+        protected string EscapeQuotes(string text)
+        {
+            return text.Replace("'", "\\'").Replace("\"", "\\\"");
+        }
+
+        protected string UnescapeQuotes(string text)
+        {
+            return text.Replace("\\'", "'").Replace("\\\"", "\"");
+        }
+
         protected abstract string GetFileText();
 
         private string GetBackupFilePath()
@@ -84,13 +71,6 @@ namespace Com.Apcurium.Resxible.Localization
                 DateTime.Now,
                 Path.GetExtension(_filePath),
                 Path.GetDirectoryName(_filePath));
-        }
-
-        private void LoadHtmlEntities()
-        {
-            _htmlEntities.Add(new HtmlEntity { Normal = "&", Encoded = "&amp;" }); // must be done first
-            _htmlEntities.Add(new HtmlEntity { Normal = "<", Encoded = "&lt;" });
-            _htmlEntities.Add(new HtmlEntity { Normal = ">", Encoded = "&gt;" });
         }
 
         private void CreateFileIfMissing()
